pyproject.toml
[project]
name = "til"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.16.2",
    "bcrypt==3.2.0",
    "dependency-injector>=4.48.1",
    "fastapi>=0.115.13",
    "mysqlclient>=2.2.7",
    "passlib[bcrypt]>=1.7.4",
    "py-ulid>=1.0.3",
    "sqlalchemy>=2.0.41",
    "uvicorn>=0.34.3",
]
